refactor: move trail to componentTrail
