+++ Source/Engine.vcxproj.filters
@@ -603,9 +603,6 @@
     <ClCompile Include="..\Game\Scripts\AnimatedTexture.cpp">
       <Filter>UserScripts\Graphics</Filter>
     </ClCompile>
-    <ClCompile Include="DataModels\Trail\Trail.cpp">
-      <Filter>DataModels\Trail</Filter>
-    </ClCompile>
     <ClCompile Include="DataModels\Components\ComponentTrail.cpp">
       <Filter>DataModels\Components</Filter>
     </ClCompile>
@@ -1252,9 +1249,6 @@
     <ClInclude Include="..\Game\Scripts\AnimatedTexture.h">
       <Filter>UserScripts\Graphics</Filter>
     </ClInclude>
-    <ClInclude Include="DataModels\Trail\Trail.h">
-      <Filter>DataModels\Trail</Filter>
-    </ClInclude>
     <ClInclude Include="DataModels\Components\ComponentTrail.h">
       <Filter>DataModels\Components</Filter>
     </ClInclude>
@@ -1423,9 +1417,6 @@
     </Filter>
     <Filter Include="UserScripts\Graphics">
       <UniqueIdentifier>{fca70cf2-3a20-410d-9019-42e680fad56c}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="DataModels\Trail">
-      <UniqueIdentifier>{6cf0d0a9-3e22-442c-be80-c9ab80a7adc5}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
